requirements.txt
fire>=0.4.0
ipython>=8.4.0
pyzmq>=23.2.1
regex>=2022.3.15
setuptools>=58.0.4
transformers>=4.22.0
torch>=1.10.0
tqdm>=4.63.0
cpm_kernels
deepspeed>0.6.1
accelerate>=0.24.1
datasets
evalplus==0.1.7
cognitive_complexity==1.3.0
radon
seaborn
structlog
wordhoard
edist
networkx
nltk
tree-sitter==0.23.1
tree-sitter-java==0.23.1
tree-sitter-c==0.23.1
tree-sitter-rust==0.23.0

numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2

